rag-chatbot/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
chromadb>=0.4.22
anthropic>=0.18.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
